#68 Moved all tests to BDD
