+++ src/test/java/bdd/CommentParsingTest.java
@@ -1,5 +1,6 @@
 package bdd;
 
+import bdd.embedders.CommentParsingEmbedder;
 import org.jbehave.core.embedder.Embedder;
 import org.jbehave.core.io.StoryFinder;
 import org.junit.Test;
