--- src/test/java/bdd/ParsingEmbedder.java
@@ -1,46 +1,0 @@
-package bdd;
-
-import bdd.steps.ManipulationSteps;
-import bdd.steps.ParsingSteps;
-import bdd.steps.SharedSteps;
-import japa.parser.ast.CompilationUnit;
-import org.jbehave.core.configuration.Configuration;
-import org.jbehave.core.configuration.MostUsefulConfiguration;
-import org.jbehave.core.embedder.Embedder;
-import org.jbehave.core.embedder.EmbedderControls;
-import org.jbehave.core.failures.FailingUponPendingStep;
-import org.jbehave.core.io.LoadFromClasspath;
-import org.jbehave.core.reporters.Format;
-import org.jbehave.core.reporters.StoryReporterBuilder;
-import org.jbehave.core.steps.InjectableStepsFactory;
-import org.jbehave.core.steps.InstanceStepsFactory;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ParsingEmbedder extends Embedder {
-
-    @Override
-    public EmbedderControls embedderControls() {
-        return new EmbedderControls().doIgnoreFailureInStories(false).doIgnoreFailureInView(false);
-    }
-
-    @Override
-    public Configuration configuration() {
-        return new MostUsefulConfiguration()
-                .useStoryLoader(new LoadFromClasspath(this.getClass()))
-                .usePendingStepStrategy(new FailingUponPendingStep())
-                .useStoryReporterBuilder(new StoryReporterBuilder()
-                        .withDefaultFormats()
-                        .withFormats(Format.CONSOLE, Format.HTML));
-    }
-
-    @Override
-    public InjectableStepsFactory stepsFactory() {
-        Map<String, Object> state = new HashMap<>();
-
-        return new InstanceStepsFactory(configuration(),
-                new SharedSteps(state),
-                new ParsingSteps(state));
-    }
-}
