+++ src/test/java/bdd/embedders/BasicJapaEmbedder.java
@@ -1,6 +1,5 @@
-package bdd;
+package bdd.embedders;
 
-import bdd.steps.CommentParsingSteps;
 import org.jbehave.core.configuration.Configuration;
 import org.jbehave.core.configuration.MostUsefulConfiguration;
 import org.jbehave.core.embedder.Embedder;
@@ -9,10 +8,12 @@ import org.jbehave.core.failures.FailingUponPendingStep;
 import org.jbehave.core.io.LoadFromClasspath;
 import org.jbehave.core.reporters.Format;
 import org.jbehave.core.reporters.StoryReporterBuilder;
-import org.jbehave.core.steps.InjectableStepsFactory;
-import org.jbehave.core.steps.InstanceStepsFactory;
 
-public class CommentParsingEmbedder extends Embedder {
+
+/**
+ * Class provides basic configuration for other embedders to reuse
+ */
+public abstract class BasicJapaEmbedder extends Embedder {
 
     @Override
     public EmbedderControls embedderControls() {
@@ -24,17 +25,10 @@ public class CommentParsingEmbedder extends Embedder {
         return new MostUsefulConfiguration()
                 // where to find the stories
                 .useStoryLoader(new LoadFromClasspath(this.getClass()))
-                // Fails if Steps are not implemented
+                        // Fails if Steps are not implemented
                 .usePendingStepStrategy(new FailingUponPendingStep())
-                // CONSOLE and HTML reporting
+                        // CONSOLE and HTML reporting
                 .useStoryReporterBuilder(new StoryReporterBuilder().withDefaultFormats()
                         .withFormats(Format.CONSOLE, Format.HTML));
     }
-
-    @Override
-    public InjectableStepsFactory stepsFactory() {
-        // varargs, can have more than one steps classes
-        return new InstanceStepsFactory(configuration(), new CommentParsingSteps());
-    }
-
 }
